add possibility to execute #toArray() with mrself/property
Initially, #toArray() used mrself/sync. But as this package is too heavy
and in most cases there is no need to use its all features, it was
decided to simplify converting entity to array by default.

There are a lot of changed files as new packages had to be installed
and accordingly some tests and other classes had to be rebuilt.